requirements.txt
streamlit
python-dotenv
httpx
pypdf2
